extract checkCodes, add test
