+++ modules/games/commands/checkCodes.ts
@@ -1,20 +1,5 @@
-import allMeta from "../dist/meta";
+import { checkCodes } from "./helpers/checkCodes";
 
-const seenCodes: Record<string, boolean> = {};
-
-for (const [gameId, meta] of Object.entries(allMeta)) {
-  const { code } = meta;
-  if (!code) {
-    throw new Error(`Game ${gameId} has no code!`);
-  }
-  if (seenCodes[code]) {
-    throw new Error(
-      `Game ${gameId} has duplicate code! Seen codes are: ${Object.keys(
-        seenCodes
-      ).join(" ")}`
-    );
-  }
-  seenCodes[code] = true;
-}
+checkCodes();
 
 console.log("All games have a unique code");
