+++ modules/games/test/uniqueCodes.test.ts
@@ -1,0 +1,7 @@
+import { checkCodes } from "../commands/helpers/checkCodes";
+
+describe("the game definitions", () => {
+  test("all have unique codes", () => {
+    expect(checkCodes).not.toThrow();
+  });
+});
